SDK: Database methods for getting filtered games #3122
