+++ source/Playnite/API/DatabaseAPI.cs
@@ -106,5 +106,15 @@ namespace Playnite.API
         {
             database.EndBufferUpdate();
         }
+
+        public bool GetGameMatchesFilter(Game game, FilterPresetSettings filterSettings)
+        {
+            return database.GetGameMatchesFilter(game, filterSettings);
+        }
+
+        public IEnumerable<Game> GetFilteredGames(FilterPresetSettings filterSettings)
+        {
+            return database.GetFilteredGames(filterSettings);
+        }
     }
 }
